updated searchRoutes to be cardRoutesand the logic to be more readable and easier to maintain for future methods that need to get added on the updated cardController
